restrict access of server side script to clear reset key

+++ clear_reset_keys.php
@@ -1,5 +1,6 @@
 <?php
-    debug_backtrace() || die("Death");
+    if (php_sapi_name() != 'cli')
+        die("No direct access, bro");
     require_once 'user_admin/php/DB_connect/db_utility.php';
 
     $deleteQuery = 'delete from ResetPassword';
